--- test/manual_tuf_repo/manifests/pvc.yaml
@@ -1,11 +1,0 @@
-apiVersion: v1
-kind: PersistentVolumeClaim
-metadata:
-  name: tuf
-  namespace: NEW_TEST_NAMESPACE
-spec:
-  accessModes:
-    - ReadWriteOnce
-  resources:
-    requests:
-      storage: 100Mi
