--- test/manual_tuf_repo/manifests/securesign.yaml
@@ -1,136 +1,0 @@
-apiVersion: rhtas.redhat.com/v1alpha1
-kind: Securesign
-metadata:
-  name: securesign-sample
-  labels:
-    app.kubernetes.io/instance: securesign-sample
-    app.kubernetes.io/name: securesign-sample
-    app.kubernetes.io/part-of: trusted-artifact-signer
-  annotations:
-    rhtas.redhat.com/metrics: 'true'
-  namespace: NEW_TEST_NAMESPACE
-spec:
-  ctlog:
-    monitoring:
-      enabled: true
-    privateKeyPasswordRef:
-      name: ctfe-secret
-      key: password
-    privateKeyRef:
-      name: ctfe-secret
-      key: private
-    publicKeyRef:
-      name: ctfe-secret
-      key: public
-    rootCertificates:
-      - key: cert
-        name: fulcio-secret
-    trillian:
-      port: 8091
-  fulcio:
-    certificate:
-      caRef:
-        key: cert
-        name: fulcio-secret
-      privateKeyPasswordRef:
-        name: fulcio-secret
-        key: password
-      privateKeyRef:
-        key: private
-        name: fulcio-secret
-      commonName: fulcio.hostname
-      organizationEmail: jdoe@redhat.com
-      organizationName: Red Hat
-    config:
-      OIDCIssuers:
-        - ClientID: trusted-artifact-signer
-          Issuer: OIDC_ISSUER_URL
-          IssuerURL: OIDC_ISSUER_URL
-          Type: email
-    ctlog:
-      port: 80
-      prefix: trusted-artifact-signer
-    externalAccess:
-      enabled: true
-    monitoring:
-      enabled: true
-  rekor:
-    monitoring:
-      enabled: true
-    signer:
-      keyRef:
-        key: private
-        name: rekor-secret
-      kms: secret
-    trillian:
-      port: 8091
-    externalAccess:
-      enabled: true
-    backFillRedis:
-      enabled: true
-      schedule: 0 0 * * *
-    rekorSearchUI:
-      enabled: true
-    pvc:
-      accessModes:
-        - ReadWriteOnce
-      retain: true
-      size: 5Gi
-  trillian:
-    database:
-      create: true
-      pvc:
-        accessModes:
-          - ReadWriteOnce
-        retain: true
-        size: 5Gi
-    monitoring:
-      enabled: true
-  tsa:
-    externalAccess:
-      enabled: true
-    monitoring:
-      enabled: true
-    ntpMonitoring:
-      enabled: true
-    signer:
-      certificateChain:
-        certificateChainRef:
-          key: certificateChain
-          name: tsa-secret
-        intermediateCA:
-          - commonName: tsa.hostname
-            organizationEmail: jdoe@redhat.com
-            organizationName: Red Hat
-        leafCA:
-          commonName: tsa.hostname
-          organizationEmail: jdoe@redhat.com
-          organizationName: Red Hat
-        rootCA:
-          commonName: tsa.hostname
-          organizationEmail: jdoe@redhat.com
-          organizationName: Red Hat
-      file:
-        passwordRef:
-          name: tsa-secret
-          key: leafPrivateKeyPassword
-        privateKeyRef:
-          key: leafPrivateKey
-          name: tsa-secret
-  tuf:
-    externalAccess:
-      enabled: true
-    keys:
-      - name: rekor.pub
-      - name: ctfe.pub
-      - name: fulcio_v1.crt.pem
-      - name: tsa.certchain.pem
-    port: 80
-    pvc:
-      accessModes:
-        - ReadWriteOnce
-      retain: true
-      size: 100Mi
-      name: tuf
-    rootKeySecretRef:
-      name: tuf-root-keys
