--- test/manual_tuf_repo/targets/ctfe-password.pem
@@ -1,1 +1,0 @@
-xDsUCJW2
