--- test/manual_tuf_repo/targets/ctfe_private_key.pem
@@ -1,8 +1,0 @@
------BEGIN EC PRIVATE KEY-----
-Proc-Type: 4,ENCRYPTED
-DEK-Info: AES-256-CBC,785981b778f883c9e15638a4b4c95af3
-
-WUzRJkSACnvoWX30qRx3OanALhw6k96GJPBPm50AJlanZTAfPavM++AkkaKzOuv2
-/0GihrNE0uK6nhpnIre962Rerlmm+d3rx0r/jxVeLiiOYrfugQ8kfnx1++9uYlvI
-hTb42zyoo4KtWO0UM6csbCyY0PrjlMGKYsvgLEeT8dI=
------END EC PRIVATE KEY-----
