--- test/manual_tuf_repo/targets/fulcio-password
@@ -1,1 +1,0 @@
-!nw@4z#N
