--- test/manual_tuf_repo/targets/fulcio_private_key.pem
@@ -1,9 +1,0 @@
------BEGIN EC PRIVATE KEY-----
-Proc-Type: 4,ENCRYPTED
-DEK-Info: AES-256-CBC,5282f0b5e37ea033b684ec066ed07020
-
-dJA/3I6kE1Q9Q5QrrCwwBIDhHSF9Du5eB3k199IWLu9AC98ke3akFLbLKE4smsOP
-Ug65Sz1JBD7MQ25xzpTyRaXrFjejtnva7smM7+a/ZqKzAWA2MJmz5mXud9eeExqC
-+bCsKlPMUwXwceqCMf/NHGHG6JuFyrr8qwgsLbOxxJDwyeU6QeSk8W5L0hq6CfuU
-LCzobKwKHtAmEMseqEabyy1I6qwl3O8qhq6O43HzYIQ=
------END EC PRIVATE KEY-----
