--- test/manual_tuf_repo/targets/rekor_private_key.pem
@@ -1,6 +1,0 @@
------BEGIN EC PRIVATE KEY-----
-MIGkAgEBBDDbJjOq3oT0Nw6jTM4w97C1Dn4DyJpWdRkPbaFHlOJjwOFgcz8grYL2
-1lOxKGFnSsqgBwYFK4EEACKhZANiAAQPDZz8LbswZsKudCPDN9+XyxbCcqeGPBrD
-XfITPn0HwYUZEmvYWIEAJ0RVpMDmzPBPDp+IH3m+od5eX7pBG5Njcpkli6Mm7ps/
-w9le1RBAOv4sm24ROuG66RDqQRQ9gaM=
------END EC PRIVATE KEY-----
