--- test/manual_tuf_repo/targets/tsa-leaf-password
@@ -1,1 +1,0 @@
-h#MQYG$9
