--- test/manual_tuf_repo/targets/tsa_leaf_private_key.pem
@@ -1,9 +1,0 @@
------BEGIN EC PRIVATE KEY-----
-Proc-Type: 4,ENCRYPTED
-DEK-Info: AES-256-CBC,e2b7891a5970dec2dcd4cc505200e5a6
-
-G36Bt3M0cp977DNERbCpvC87W9U3czT76LPtnGVN3cFNi9hNUs9L63vk3giv8US+
-264SToikuR/9L9fTIbhnkK557HoWc6VcIIkHlTHc0qyQrPzvxVnX0sHOu6WO3C72
-BJzllk8MbYQmg2ylKtb9rFn36CP8qiffvzlz+JnBmHJdQbNvcPuKomElYxMKQG2Q
-spDaSkAZPQXfOFYB0mERmxUm2k1oO7nPZ7pyHtLfA00=
------END EC PRIVATE KEY-----
