+++ test/tuftool/tuftool_suite_init_test.go
@@ -1,4 +1,4 @@
-package manualtuf
+package tuftool
 
 import (
 	"testing"
